requirements.txt
aiohttp==3.7.4.post0
astroid==2.11.5
async-timeout==3.0.1
attrs==21.4.0
black==22.3.0
certifi==2022.5.18.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
dill==0.3.5.1
discord.py==1.7.3
et-xmlfile==1.1.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
multidict==6.0.2
mypy-extensions==0.4.3
numpy==1.22.4
openai==0.19.0
openpyxl==3.0.10
pandas==1.4.2
pandas-stubs==1.2.0.58
pathspec==0.9.0
platformdirs==2.5.2
pylint==2.13.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
tomli==2.0.1
tqdm==4.64.0
typing-extensions==4.2.0
urllib3==1.26.9
wrapt==1.14.1
yarl==1.7.2
